[GUI] Added comments, WIP mods to BeginTextUpdate() to automatically place the cursor correctly (needs testing) [MATHS] Added vector == operator and a NullVector as an argument to test against [MISC] Minor bug fixes / code cleanup, added batch file for creation of dll with debugger assertions

+++ prepare_release_files.bat
@@ -1,5 +1,11 @@
 @echo off
 
+if %1.==. ( 
+	echo:
+	echo Usage: %0 [build.bat without the path]
+	exit /b 
+)
+
 if exist release (rmdir release /s /q)
 mkdir release
 mkdir release\bin
@@ -7,7 +13,7 @@ mkdir release\source
 
 pushd source\
 if exist build (del build /q)
-call build_dll_debug.bat
+call %1
 popd
 
 echo:
